+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 ## 0.9.3 (not released yet)
 
 * Added [callback](lib/Widget/Callback.php) widget to handle WeChat(Weixin) callback message
-* Added `overwirte` option for [upload](lib/Widget/Upload.php) widget
+* Added `overwrite` option for [upload](lib/Widget/Upload.php) widget
 * Added support for upload file without extension
 * Added `getMulti` and `setMulti` method for cache widgets
 * Added code completion supports for widgets, refs [AbstractWidget](lib/Widget/AbstractWidget.php)
@@ -13,7 +13,7 @@ Changelog
 * Removed `inMethod` widget
 * Added new API documentation: http://twinh.github.io/widget/
 * Added `getJoinedMessage` method for validators, refs #52
-* Simplfied [cookie](lib/Widget/Cookie.php) widget options, refs fa88083a742f7aa7e8d6d1829a34e4ca853fb50a
+* Simplified [cookie](lib/Widget/Cookie.php) widget options, refs fa88083a742f7aa7e8d6d1829a34e4ca853fb50a
 * Removed `inGet`, `inPost`, `inAjax` widget, use [request](lib/Widget/Request.php) widget instead
 * Merged `sort` and `attr` widgets in to [arr](lib/Widget/Arr.php) widget
 * Refactored [dbCache](lib/Widget/DbCache.php) widget
@@ -36,7 +36,7 @@ Changelog
 * Released first beta version
 * Added unit test and fixed lots of error for all widgets
 * Added validation component
-* Added fisrt version of API documentation
+* Added first version of API documentation
 
 ## 2012-08-30
 * Moved to GitHub